+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [unreleased]
+- call unBondNodes in the protocol, which unbonds the nodes without the tokens
+
 ## [0.5.5]
 - unbond doesn't run out of gas, it simply interrupts its execution